+++ spaghetti/src/utils/fetch.js
@@ -1,6 +1,11 @@
-import axios from "axios";
+import axios from 'axios'
 
-export const fetchDataListAndPagination = async ({ endpoints, page = 1, take = 20, limit = 10 }) => {
+export const fetchDataListAndPagination = async ({
+	endpoints,
+	page = 1,
+	take = 20,
+	limit = 10
+}) => {
 	const result = { dataList: {}, pagination: {} }
 	try {
 		const response = await axios.get(endpoints, {
@@ -11,8 +16,7 @@ export const fetchDataListAndPagination = async ({ endpoints, page = 1, take = 2
 		result.pagination = pagination
 		result.dataList = dataList
 		return result
-	}
-	catch {
+	} catch {
 		return result
 	}
 }
@@ -22,7 +26,7 @@ export const fetchPostDetailById = async (postId) => {
 	try {
 		const response = await axios.get(`/api/post/${postId}`)
 		return response.data
-	} catch { 
+	} catch {
 		return result
 	}
 }
